updated batch cloning to use the new form. Also refactored so there isn't 3 copies of the same form component anymore
